Redesign sched schema
This revision introduces *pattern* and *template* models that enable 
handling network itineraries and operation schedules separately. 
Generally, itineraries of route variants on along the network are 
relatively stable, often including only the two directions, sometimes 
additional variants e.g. for depot divertions. These "patterns" provide 
a meaningful common basis for many kinds of analyses. On the contrary, 
operating times of a route can vary a lot, depending on the time, so it 
is reasonable to separate them from the itinerary model.