Include linear loc priority
Partly related to the previous commit by which we kind of opened the 
areas around segment ends for potential points: we do not in fact want 
to do that, at least anywhere else than the first and last segment of 
ptid. Prioritizing segment matches with rel loc other than 0 or 1 does 
effectively the same thing but faster.