Move old DDL and run scripts to old_ddl/
Also add a readme about the purpose of the directory. I think it's 
easier to keep these in the project dir for reference than to look up 
old files from last year from the git history, as we still want to 
implement some things nearly the same way (but differently enough to 
create new, clean script files). This should be deleted when we're 
(nearly) ready with the new database model.