Clean up `db/` from old files (#47)
* Remove db/explore

These files have only been for my own "sketching" and they are not relevant for production or testing.

* Remove db/patches

These have been dataset-specific fixes and not relevant with new data 
any more.

* Move old DDL and run scripts to old_ddl/

Also add a readme about the purpose of the directory. I think it's 
easier to keep these in the project dir for reference than to look up 
old files from last year from the git history, as we still want to 
implement some things nearly the same way (but differently enough to 
create new, clean script files). This should be deleted when we're 
(nearly) ready with the new database model.

* Remove old files

* Move executable files from db/db_test to db/

* Remove config/

This was only related to the OSM data imports that we no longer use.