Find links with distance & side
using lateral join where the stop mode must be allowed in the link modes 
(note the ARRAY[] because of the <@ operator), and the stop _must_ be 
within X meters to the right of the link. Only these candidates are then 
sorted and the link closest to the stop is then selected.