Add via_nodes attribute
I think this will be more useful than passing via_nodes as argument to a 
function that runs the routing and creates link_on_section rows. This 
way the user can define via_nodes for many sections beforehand and then 
run the routing as a batch process. Also, this way the source data for 
how the routing was made can be preserved.