Add "expanded" view for stop_on_route
This view connects route versions, route version stops, stop links and 
nodes. It works both for visual inspection in QGIS and as a basis for 
listing route version via nodes based on ordered stops.

All the fields are not necessary, on the other hand we might want to add 
new fields later...