Add upsert_link_on_route procedures
Closes #58. Now we are able to automatically create shortest path 
link_on_route sequences using via nodes from stop_on_route and 
manual_vianode_on_route. Can be run for a single route version id passed 
by the user or as batch procedure for all route versions available. Like 
with analysis sections, looping over stuff and re-querying it during 
each step is not efficient and elegant, but again we are not dealing 
with big data in this case.