Change smallint link_dir to boolean link_reversed
This is a clearer way to model the link traversal direction as we only
need two values (+ NULL) anyway.
We can also easily cast this to 0 / 1 int to use in calculations that
determine whether relative distance values along a link should be
reversed or not. Assume x is a relative location between 0.0 and 1.0:

  link_reversed is false -> 0, result = abs(0 - x) = x
  link_reversed is true -> 1, result = abs(1 - x) = 1 - x

I had accidentally thought that values (-1, 1) would handle this better,
but this was not the case. Therefore I am making the change from
smallint to boolean and hopefully making the data model a bit easier
to understand at the same time.