Change bind source to a fixed directory
env_file will import the env variables into the container, but the 
substitution will not work here - IMPORT_DATA_DIR seems to be whatever 
is defined in .env. For ./test.sh runs, we are probably always going to 
use example_data/ contents, so we can as well fix the directory here.