Create obs.hfp_point model
and add corresponding example data and import command. obs.hfp_point 
will be a large table, storing thousands of HFP points for each journey, 
and therefore we partition it with Timescale. At this point, time and 
hash partition sizes are somewhat arbitrary, and they should be 
optimized later if performance problems occur.