Create initial model and algorithm for HFP points matched to links
See the note in 060_functions_obs.sql. Per-link aggregates and other 
space-time measures will produce spurious results before we develop this 
algorithm so that it tries to prefer link candidates that result in a 
consistent path of projected HFP observations that do not move back and 
forth along the route (unless there really isn't another option).