Create view of sections with i and j nodes and i-j line geom
This view primarily aims at easier interactive modifications of analysis 
sections e.g. in QGIS. A significant drawback is that nodes are fixed to 
i (start) and j (end) nodes; we could later improve this so that the 
linestring geometry consists of _all_ via nodes in the correct order, 
and thus via nodes could be altered by modifying the linestring 
geometry.