Add obs.halt_on_journey table
This table will be the basis for analyses where we are interested in how 
non-movement time of journeys is spent: when, how long and how often, 
and by joining obs.point_on_link on (jrnid, tst) also "where". 
Calculating the halt events as single-row entries from points_on_link is 
a rather complex operation, and therefore we materialize them into a 
separate table. We do not want to use a materialized view (at least for 
now) because refreshing it would re-calculate all data, while in reality 
we are likely to do incremental imports and updates instead (e.g. by 
route, dir and oday).