Set link_id to NULL on link delete
This enables flexible initial modifications to the network.
Stops have their own geometries, and link references are, in a way,
"extra information" on top of that.
Note that other link ref info is not set to NULl on the stop record
when the link is deleted, but this mismatch should be noted by
validations and at least overwritten when a new link is added
and the stop-link ref info is updated.