Update parent and link fkey actions
The id of the parent (route version, section) is arbitrary
and thus the updates should go seamlessly downstream.
Deleting the parent object naturally results in deleting the
related link references.

link_id is an essential reference as well and cannot be set to NULL,
therefore the record is deleted if the link is deleted.
There can be cases where the link_id gets a new value (e.g.,
upon splitting a link), so it makes sense to cascade this update.