Update journey and link fkey actions
If a journey is deleted, its child objects must be deleted as well:
journeys and observations are imported (and thus deleted) together.
Note that cascading referenced route_ver_id update or deletion is
restricted: route_ver_id is dependent on the route version validity
time and changes to it should be done with care, controlling possible
referred journeys at the same time. Deletions are restricted to avoid
accidental loss of large amount of data when editing the network.

link_id references are "essential" ones, so it makes no sense to keep
the record if the referred link is deleted. Updating the link_id
also updates it here, but note that the user must check otherwise if
the data is still consistent: e.g., after splitting a link (and renaming
link_ids therefore) we should update related obs data anyway.